Alternative parts for the wheel and buffer front.
The wheel here has an extra hole on the other edge to accommodate loading the filament from right to left without possible tangles occurring (due to the inlet/outlet holes of the buffer top being offset front to back.)  This was produced in Fusion360 using the original ERCP model, thickened to 14.8mm to match the Pika wheel, and holes/notches added.  (The .step file is included.)

The array-front(merged) file is a simple merging of "array-front-screen" and "array-front" STL's.  (The screen was thickened by 0.4mm to create a solid object.)   If both parts are printed in the same color, printing a single object is significantly faster than printing two objects (that each take an entire build plate.)  This merge was done in tinkercad, so there's no .step file to include.